Refactor config and variable logic into multiple, smaller files
