+++ util/collections.go
@@ -1,5 +1,7 @@
 package util
 
+import "fmt"
+
 // Merge all the maps into one. Sadly, Go has no generics, so this is only defined for string to interface maps.
 func MergeMaps(maps ... map[string]interface{}) map[string]interface{} {
 	out := map[string]interface{}{}
@@ -22,4 +24,31 @@ func ListContains(needle string, haystack []string) bool {
 	}
 
 	return false
+}
+
+// Convert a generic list to a list of strings
+func ToStringList(genericList []interface{}) []string {
+	stringList := []string{}
+
+	for _, value := range genericList {
+		stringList = append(stringList, ToString(value))
+	}
+
+	return stringList
+}
+
+// Convert a generic map to a map from string to string
+func ToStringMap(genericMap map[interface{}]interface{}) map[string]string {
+	stringMap := map[string]string{}
+
+	for key, value := range genericMap {
+		stringMap[ToString(key)] = ToString(value)
+	}
+
+	return stringMap
+}
+
+// Convert a single value to its string representation
+func ToString(value interface{}) string {
+	return fmt.Sprintf("%v", value)
 }
